refactor: extract FolderTabbedPane inner classes into delegates (643→363 lines)
Extract four inner classes into package-private delegate files:
- FolderTabTitleUI: custom UI delegate for tab title buttons
- JFolderTabTitle: custom JToggleButton for tab titles
- FolderTitlesPanel: panel holding/painting tab titles with curved shapes
- TabScrollListener: mouse listener for drag-scrolling tab titles

Add FolderTabbedPaneApiTest characterization test verifying the public API
is preserved (superclass, type parameters, public methods, constructor).

Closes #628
