+++ core/croquet/src/main/java/org/lgna/croquet/views/FolderTabbedPane.java
@@ -52,129 +52,20 @@ import org.lgna.croquet.triggers.Trigger;
 
 import javax.swing.*;
 import javax.swing.border.Border;
-import javax.swing.plaf.basic.BasicToggleButtonUI;
 import java.awt.*;
-import java.awt.event.*;
-import java.awt.geom.GeneralPath;
+import java.awt.event.ActionListener;
 import java.util.UUID;
 
 /**
- * The folder tabbed pane ecosystem (there are so many classes in just this one file!) appears to be some slightly
- * modified version of the swing JTabbedPane.  It would be worth looking into if there is a simpler way to get whatever
- * non-standard behavior we need (or at least document in a comment WHY we needed to reinvent this)
+ * A tabbed pane with folder-style curved tabs. Delegates tab painting to
+ * {@link FolderTitlesPanel}, tab title UI to {@link FolderTabTitleUI} /
+ * {@link JFolderTabTitle}, and drag-scroll behaviour to {@link TabScrollListener}.
  *
  * @author Dennis Cosgrove
  */
 public class FolderTabbedPane<E extends TabComposite<?>> extends CardBasedTabbedPane<E> {
-  private static final int TRAILING_TAB_PAD = 32;
+  static final int TRAILING_TAB_PAD = 32;
   private static final int OUTLINE_THICKNESS = 1;
-
-  private static class FolderTabTitleUI extends BasicToggleButtonUI {
-    @Override
-    public Dimension getPreferredSize(JComponent c) {
-      javax.swing.AbstractButton button = (javax.swing.AbstractButton) c;
-      Font font = button.getFont();
-      FontMetrics fm = button.getFontMetrics(font);
-      String text = button.getText();
-      Icon icon = button.getIcon();
-      Dimension size;
-      if (icon != null) {
-        int verticalAlignment = button.getVerticalAlignment();
-        int horizontalAlignment = button.getHorizontalAlignment();
-        int verticalTextPosition = button.getVerticalTextPosition();
-        int horizontalTextPosition = button.getHorizontalTextPosition();
-        Rectangle viewR = new Rectangle(Short.MAX_VALUE, Short.MAX_VALUE);
-        Rectangle iconR = new Rectangle();
-        Rectangle textR = new Rectangle();
-        int textIconGap = button.getIconTextGap();
-        SwingUtilities.layoutCompoundLabel(c, fm, text, icon, verticalAlignment, horizontalAlignment, verticalTextPosition, horizontalTextPosition, viewR, iconR, textR, textIconGap);
-
-        size = iconR.union(textR).getSize();
-      } else {
-        size = fm.getStringBounds(text, button.getGraphics()).getBounds().getSize();
-      }
-
-      Insets insets = button.getInsets();
-      size.width += insets.left + insets.right;
-      size.height += insets.top + insets.bottom;
-
-      if (button.getComponentCount() > 0) {
-        for (Component component : button.getComponents()) {
-          size.width += 4;
-          size.width += component.getPreferredSize().width;
-        }
-      }
-
-      return size;
-    }
-
-    @Override
-    public void paint(Graphics g, JComponent c) {
-      javax.swing.AbstractButton button = (javax.swing.AbstractButton) c;
-      Icon icon = button.getIcon();
-      if (icon != null) {
-        super.paint(g, c);
-      } else {
-        String text = button.getText();
-        Insets insets = button.getInsets();
-        int x = insets.left;
-        if (!button.getComponentOrientation().isLeftToRight()) {
-          for (Component component : button.getComponents()) {
-            x += component.getPreferredSize().width;
-            x += 4;
-          }
-        }
-        Color outlineColor = ColorUtilities.scaleHSB(button.getBackground(), 1, 4.8, .74);
-        g.setColor(button.isSelected() ? UIManager.getColor("TabbedPane.foreground") :  button.getModel().isRollover() ? UIManager.getColor("TabbedPane.disabledForeground") : outlineColor);
-        g.drawString(text, x, button.getBaseline(c.getWidth(), c.getHeight()));
-      }
-    }
-  }
-
-  private static class JFolderTabTitle extends JToggleButton {
-    private final ItemListener itemListener = e -> JFolderTabTitle.this.revalidate();
-
-    public JFolderTabTitle() {
-      this.setBorder(BorderFactory.createEmptyBorder(4, 4, 4, 8));
-      this.setAlignmentX(Component.LEFT_ALIGNMENT);
-      this.setAlignmentY(Component.BOTTOM_ALIGNMENT);
-      this.setHorizontalTextPosition(SwingConstants.LEADING);
-      this.setHorizontalAlignment(SwingConstants.LEADING);
-      this.setLayout(new SpringLayout());
-    }
-
-    @Override
-    public boolean isOpaque() {
-      return false;
-    }
-
-    @Override
-    public void updateUI() {
-      this.setUI(new FolderTabTitleUI());
-    }
-
-    @Override
-    public void addNotify() {
-      super.addNotify();
-      this.getModel().addItemListener(this.itemListener);
-    }
-
-    @Override
-    public void removeNotify() {
-      this.getModel().removeItemListener(this.itemListener);
-      super.removeNotify();
-    }
-
-    @Override
-    public void repaint() {
-      Container parent = this.getParent();
-      if (parent != null) {
-        parent.repaint(this.getX(), this.getY(), this.getWidth() + TRAILING_TAB_PAD, this.getHeight());
-      } else {
-        super.repaint();
-      }
-    }
-  }
 
   private class FolderTabTitle extends BooleanStateButton<javax.swing.AbstractButton> {
     private final JButton closeButton;
@@ -218,136 +109,15 @@ public class FolderTabbedPane<E extends TabComposite<?>> extends CardBasedTabbed
     }
   }
 
-  protected static class TitlesPanel extends LineAxisPanel {
-    private static final int NORTH_AREA_PAD = 1;
-
-    protected static class JTitlesPanel extends JPanel {
-      @Override
-      public Dimension getPreferredSize() {
-        Dimension rv = super.getPreferredSize();
-        rv.width += TRAILING_TAB_PAD;
-        return rv;
-      }
-
-      private GeneralPath addToPath(GeneralPath rv, float x, float y, float width, float height) {
-        float a = height * 0.25f;
-
-        float xStart;
-        float xEnd;
-        float xA;
-        float tabPad;
-        if (this.getComponentOrientation().isLeftToRight()) {
-          xStart = x;
-          xEnd = (x + width) - 1;
-          tabPad = TRAILING_TAB_PAD;
-          xA = xStart + a;
-        } else {
-          xStart = (x + width) - 1;
-          xEnd = x;
-          tabPad = -TRAILING_TAB_PAD;
-          xA = xStart - a;
-        }
-
-        float xCurve0 = xEnd - (tabPad / 2);
-        float xCurve1 = xEnd + tabPad;
-        float cx0 = xCurve0 + (tabPad * 0.75f);
-        float cx1 = xCurve0;
-
-        float y0 = y + NORTH_AREA_PAD;
-        float y1 = y + height + 1; // + this.contentBorderInsets.top;
-        float cy0 = y0;
-        float cy1 = y1;
-
-        float yA = y + a;
-
-        rv.moveTo(xCurve1, y1);
-
-        rv.lineTo(xCurve1, y1 - 1);
-        rv.curveTo(cx1, cy1, cx0, cy0, xCurve0, y0);
-        rv.lineTo(xA, y0);
-        rv.quadTo(xStart, y0, xStart, yA);
-        rv.lineTo(xStart, y1);
-
-        return rv;
-      }
-
-      private void paintTab(Graphics2D g2, javax.swing.AbstractButton button) {
-        Color prevColor = g2.getColor();
-        Shape prevClip = g2.getClip();
-
-        try {
-          int x = button.getX();
-          int y = button.getY();
-          int width = button.getWidth();
-          int height = button.getHeight();
-
-          Color color = button.getBackground();
-          Color outlineColor = ColorUtilities.scaleHSB(color, 1, 4.8, .74);
-
-          if (button.isSelected()) {
-            // draw one more pixel down to connect with the panel outline
-            Rectangle bounds = prevClip.getBounds();
-            bounds.height += 1;
-            g2.setClip(bounds);
-          } else {
-            color = button.getModel().isRollover() ? color : ColorUtilities.scaleHSB(color, 1, .5, 1);
-          }
-          g2.setColor(color);
-
-          GeneralPath path = addToPath(new GeneralPath(), x, y, width, height);
-
-          // draw the background before the outline
-          g2.fill(path);
-          g2.setColor(outlineColor);
-          g2.draw(path);
-        } finally {
-          g2.setColor(prevColor);
-        }
-      }
-
-      @Override
-      protected void paintChildren(Graphics g) {
-        Graphics2D g2 = (Graphics2D) g;
-        Object prevAntialiasing = g2.getRenderingHint(RenderingHints.KEY_ANTIALIASING);
-        g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
-        javax.swing.AbstractButton selectedButton = null;
-        Component[] components = this.getComponents();
-        final int N = components.length;
-        for (int i = 0; i < N; i++) {
-          Component component = components[N - 1 - i];
-          if (component instanceof javax.swing.AbstractButton button) {
-            if (button.isSelected()) {
-              selectedButton = button;
-            } else {
-              this.paintTab(g2, button);
-            }
-          }
-        }
-        // paint selected button last so that it shows up on top
-        if (selectedButton != null) {
-          this.paintTab(g2, selectedButton);
-        }
-        super.paintChildren(g2);
-        g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, prevAntialiasing == null ? RenderingHints.VALUE_ANTIALIAS_DEFAULT : prevAntialiasing);
-      }
-    }
-
-    @Override
-    protected JPanel createJPanel() {
-      return new JTitlesPanel();
-    }
-  }
-
-  private final TitlesPanel titlesPanel = this.createTitlesPanel();
+  private final FolderTitlesPanel titlesPanel = this.createTitlesPanel();
   private final ScrollPane titlesScrollPane = new ScrollPane(this.titlesPanel);
   private final BorderPanel innerHeaderPanel = new BorderPanel();
   private final BorderPanel outerHeaderPanel = new BorderPanel();
 
-  protected TitlesPanel createTitlesPanel() {
-    return new TitlesPanel();
+  protected FolderTitlesPanel createTitlesPanel() {
+    return new FolderTitlesPanel();
   }
 
-  //private java.util.Map<E, javax.swing.Action> mapItemToAction = edu.cmu.cs.dennisc.java.util.Maps.newHashMap();
   private Action getActionFor(E item) {
     Operation operation = this.getModel().getItemSelectionOperation(item);
     operation.initializeIfNecessary();
@@ -385,7 +155,6 @@ public class FolderTabbedPane<E extends TabComposite<?>> extends CardBasedTabbed
       popupMenu.show(viewController.getAwtComponent(), 0, viewController.getHeight());
     }
   }
-
   private class PopupButton extends OperationButton<JButton, Operation> {
     public PopupButton(Operation operation) {
       super(operation);
@@ -439,61 +208,13 @@ public class FolderTabbedPane<E extends TabComposite<?>> extends CardBasedTabbed
     }
   }
 
-  private class ScrollListener implements MouseListener, MouseMotionListener {
-    private Integer pressedLocationX;
-    private Integer pressedViewPositionX;
-
-    @Override
-    public void mouseClicked(MouseEvent e) {
-    }
-
-    @Override
-    public void mousePressed(MouseEvent e) {
-      this.pressedLocationX = e.getPoint().x;
-      this.pressedViewPositionX = titlesScrollPane.getAwtComponent().getViewport().getViewPosition().x;
-    }
-
-    @Override
-    public void mouseReleased(MouseEvent e) {
-      this.pressedLocationX = null;
-      this.pressedViewPositionX = null;
-    }
-
-    @Override
-    public void mouseEntered(MouseEvent e) {
-    }
-
-    @Override
-    public void mouseExited(MouseEvent e) {
-    }
-
-    @Override
-    public void mouseMoved(MouseEvent e) {
-    }
-
-    @Override
-    public void mouseDragged(MouseEvent e) {
-      if (this.pressedLocationX != null) {
-        Dimension viewSize = titlesScrollPane.getAwtComponent().getViewport().getView().getSize();
-        Rectangle viewportRect = titlesScrollPane.getAwtComponent().getViewport().getViewRect();
-
-        int xDelta = this.pressedLocationX - e.getX();
-        int value = this.pressedViewPositionX + xDelta;
-        value = Math.max(value, 0);
-        value = Math.min(value, viewSize.width - viewportRect.width);
-
-        titlesScrollPane.getAwtComponent().getViewport().setViewPosition(new Point(value, 0));
-      }
-    }
-  }
-
   public FolderTabbedPane(TabState<E, ?> model) {
     super(model);
     CardOwnerComposite cardOwner = this.getCardOwner();
     this.titlesScrollPane.setHorizontalScrollbarPolicy(ScrollPane.HorizontalScrollbarPolicy.NEVER);
     this.titlesScrollPane.setVerticalScrollbarPolicy(ScrollPane.VerticalScrollbarPolicy.NEVER);
 
-    ScrollListener scrollListener = new ScrollListener();
+    TabScrollListener scrollListener = new TabScrollListener(this.titlesScrollPane);
     this.titlesScrollPane.addMouseListener(scrollListener);
     this.titlesScrollPane.addMouseMotionListener(scrollListener);
 
@@ -542,7 +263,6 @@ public class FolderTabbedPane<E extends TabComposite<?>> extends CardBasedTabbed
     this.setInnerHeaderTrailingComponent(new PopupButton(popupOperation));
 
   }
-
   public void setHeaderLeadingComponent(SwingComponentView<?> component) {
     if (component != null) {
       component.setAlignmentY(Component.BOTTOM_ALIGNMENT);
