+++ core/croquet/src/main/java/org/lgna/croquet/views/TabScrollListener.java
@@ -1,0 +1,103 @@
+/*******************************************************************************
+ * Copyright (c) 2006, 2015, Carnegie Mellon University. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * 3. Products derived from the software may not be called "Alice", nor may
+ *    "Alice" appear in their name, without prior written permission of
+ *    Carnegie Mellon University.
+ *
+ * 4. All advertising materials mentioning features or use of this software must
+ *    display the following acknowledgement: "This product includes software
+ *    developed by Carnegie Mellon University"
+ *
+ * 5. The gallery of art assets and animations provided with this software is
+ *    contributed by Electronic Arts Inc. and may be used for personal,
+ *    non-commercial, and academic use only. Redistributions of any program
+ *    source code that utilizes The Sims 2 Assets must also retain the copyright
+ *    notice, list of conditions and the disclaimer contained in
+ *    The Alice 3.0 Art Gallery License.
+ *
+ * DISCLAIMER:
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND.
+ * ANY AND ALL EXPRESS, STATUTORY OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY,  FITNESS FOR A
+ * PARTICULAR PURPOSE, TITLE, AND NON-INFRINGEMENT ARE DISCLAIMED. IN NO EVENT
+ * SHALL THE AUTHORS, COPYRIGHT OWNERS OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+ * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, PUNITIVE OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING FROM OR OTHERWISE RELATING TO
+ * THE USE OF OR OTHER DEALINGS WITH THE SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *******************************************************************************/
+
+package org.lgna.croquet.views;
+
+import java.awt.*;
+import java.awt.event.*;
+
+/**
+ * Mouse listener that enables drag-scrolling on the tab titles scroll area.
+ * Extracted from {@link FolderTabbedPane}.
+ */
+class TabScrollListener implements MouseListener, MouseMotionListener {
+  private final ScrollPane scrollPane;
+  private Integer pressedLocationX;
+  private Integer pressedViewPositionX;
+
+  TabScrollListener(ScrollPane scrollPane) {
+    this.scrollPane = scrollPane;
+  }
+  @Override
+  public void mouseClicked(MouseEvent e) {
+  }
+
+  @Override
+  public void mousePressed(MouseEvent e) {
+    this.pressedLocationX = e.getPoint().x;
+    this.pressedViewPositionX = this.scrollPane.getAwtComponent().getViewport().getViewPosition().x;
+  }
+
+  @Override
+  public void mouseReleased(MouseEvent e) {
+    this.pressedLocationX = null;
+    this.pressedViewPositionX = null;
+  }
+
+  @Override
+  public void mouseEntered(MouseEvent e) {
+  }
+
+  @Override
+  public void mouseExited(MouseEvent e) {
+  }
+
+  @Override
+  public void mouseMoved(MouseEvent e) {
+  }
+
+  @Override
+  public void mouseDragged(MouseEvent e) {
+    if (this.pressedLocationX != null) {
+      Dimension viewSize = this.scrollPane.getAwtComponent().getViewport().getView().getSize();
+      Rectangle viewportRect = this.scrollPane.getAwtComponent().getViewport().getViewRect();
+
+      int xDelta = this.pressedLocationX - e.getX();
+      int value = this.pressedViewPositionX + xDelta;
+      value = Math.max(value, 0);
+      value = Math.min(value, viewSize.width - viewportRect.width);
+
+      this.scrollPane.getAwtComponent().getViewport().setViewPosition(new Point(value, 0));
+    }
+  }
+}
